src/lambdas/prediction/requirements.txt
python-dotenv
pycaret==3.0.4
pandas==1.5.3
psycopg2-binary
pydantic
fastapi